requirements.txt
openai>=1.0.0
python-dateutil==2.8.2
python-dotenv>=1.0.0
argparse>=1.4.0